requirements.txt.txt
streamlit
torch==2.0.1
torchvision==0.15.2
Pillow